make mono_wasm_invoke_method_bound safer with exception root
